# openapi.yaml
...
openapi: 3.1.0
info:
  title: EcomGen API
  version: 0.1.0
  description: Backend-first contract for the EcomGen desktop-first local application.
security: []
servers:
  - url: http://localhost:8787/api/v1
paths:
...
  /projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: getProject
      responses:
        "200":
          description: Project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectInput"
      responses:
        "200":
          description: Project updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      responses:
        "204":
          description: Project deleted.
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{projectId}/assets:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: listAssets
      responses:
        "200":
          description: Assets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetList"
    post:
      operationId: uploadAsset
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadAssetInput"
      responses:
        "200":
          description: Asset uploaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
...
  /projects/{projectId}/storyboard:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: getStoryboard
      responses:
        "200":
          description: Current storyboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryboardBundle"
  /projects/{projectId}/storyboard/confirm:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      operationId: confirmStoryboard
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmStoryboardInput"
      responses:
        "200":
          description: Storyboard frozen for generation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Storyboard"
        "409":
          $ref: "#/components/responses/Conflict"
  /storyboard-items/{itemId}:
    parameters:
      - $ref: "#/components/parameters/StoryboardItemId"
    patch:
      operationId: updateStoryboardItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoryboardItemInput"
      responses:
        "200":
          description: Storyboard item updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryboardItem"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteStoryboardItem
      responses:
        "204":
          description: Draft storyboard item deleted.
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{projectId}/generation-jobs:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      operationId: createGenerationJobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGenerationJobInput"
      responses:
        "202":
          description: Generation job queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJobsResponse"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    StoryboardItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Current resource state rejects this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    AssetRole:
      type: string
      enum:
        - PRODUCT_TRUTH
        - PACKAGING
        - STYLE_REFERENCE
        - LAYOUT_REFERENCE
...
    ImageResolution:
      type: string
      enum:
        - 1K
        - 2K
        - 4K
    ImageAspectRatio:
      type: string
      enum:
        - AUTO
        - 1:1
        - 2:3
        - 3:2
        - 3:4
        - 4:3
        - 4:5
        - 5:4
        - 9:16
        - 16:9
        - 21:9
...
    Asset:
      type: object
      required:
        - id
        - projectId
        - role
        - mimeType
        - createdAt
      properties:
        id:
          format: uuid
          type: string
        projectId:
          format: uuid
          type: string
        role:
          $ref: "#/components/schemas/AssetRole"
        kind:
          $ref: "#/components/schemas/UserAssetKind"
        url:
          type: string
        storagePath:
          type: string
        mimeType:
          type: string
        originalName:
          type: string
        hash:
          type: string
        width:
          anyOf:
            - type: integer
            - type: "null"
        height:
          anyOf:
            - type: integer
            - type: "null"
        createdAt:
          format: date-time
          type: string
    AssetList:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
        nextCursor:
          anyOf:
            - type: string
            - type: "null"
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                patternProperties:
                  ^(.*)$: {}
            requestId:
              format: uuid
              type: string
...
    Job:
      type: object
      required:
        - id
        - type
        - status
        - progress
        - retryable
        - createdAt
      properties:
        id:
          format: uuid
          type: string
        type:
          $ref: "#/components/schemas/JobType"
        status:
          type: string
          enum:
            - QUEUED
            - RUNNING
            - SUCCEEDED
            - FAILED
            - CANCELLED
        progress:
          minimum: 0
          maximum: 100
          type: integer
        retryable:
          type: boolean
        requestFingerprint:
          anyOf:
            - type: string
            - type: "null"
        providerId:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
        modelId:
          anyOf:
            - type: string
            - type: "null"
        estimatedCost:
          anyOf:
            - type: object
              patternProperties:
                ^(.*)$: {}
            - type: "null"
        actualCost:
          anyOf:
            - type: object
              patternProperties:
                ^(.*)$: {}
            - type: "null"
        cancelRequested:
          type: boolean
        error:
          anyOf:
            - type: object
              patternProperties:
                ^(.*)$: {}
            - type: "null"
        progressDetail:
          anyOf:
            - $ref: "#/components/schemas/JobProgressDetail"
            - type: "null"
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
...
    ModelRef:
      type: object
      required:
        - providerId
        - modelId
      properties:
        providerId:
          format: uuid
          type: string
        modelId:
          type: string
    Output:
      type: object
      required:
        - id
        - storyboardItemId
        - jobId
        - createdAt
      properties:
        id:
          format: uuid
          type: string
        storyboardItemId:
          format: uuid
          type: string
        jobId:
          format: uuid
          type: string
        candidateIndex:
          minimum: 1
          type: integer
        generationSnapshot:
          anyOf:
            - type: object
              properties:
                providerId:
                  format: uuid
                  type: string
                modelId:
                  type: string
                resolution:
                  $ref: "#/components/schemas/ImageResolution"
                aspectRatio:
                  $ref: "#/components/schemas/ImageAspectRatio"
                size:
                  type: string
                quality:
                  description: 本次请求实际下发的质量档位；不消费该参数的模型为空。
                  $ref: "#/components/schemas/ImageQuality"
                outputFormat:
                  description: 本次请求实际请求的图片格式；不消费该参数的模型为空。
                  $ref: "#/components/schemas/ImageOutputFormat"
                candidateIndex:
                  minimum: 1
                  type: integer
                revision:
                  type: string
            - type: "null"
        url:
          type: string
        storagePath:
          type: string
        parentOutputId:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
        rootOutputId:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
        editSessionId:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
        editTurnId:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
        generationBatchId:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
        createdAt:
          format: date-time
          type: string
...
    Project:
      type: object
      required:
        - id
        - name
        - platformTargets
        - targetMarket
        - copyLanguage
        - reasoningProviderId
        - reasoningModelId
        - imageProviderId
        - imageModelId
        - defaultMode
        - imageResolution
        - imageAspectRatio
        - candidatesPerType
        - webResearchEnabled
        - createdAt
        - updatedAt
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        category:
          anyOf:
            - type: string
            - type: "null"
        productDescription:
          anyOf:
            - type: string
            - type: "null"
        verifiedFacts:
          type: array
          items:
            type: string
        prohibitedClaims:
          type: array
          items:
            type: string
        brandGuidelines:
          type: object
          patternProperties:
            ^(.*)$: {}
        platformTargets:
          maxItems: 1
          type: array
          items:
            type: string
            enum:
              - TAOBAO
              - JD
              - PDD
              - DOUYIN
              - AMAZON
              - SHOPIFY
        targetMarket:
          anyOf:
            - const: CHINA_MAINLAND
              type: string
            - const: HONG_KONG
              type: string
            - const: MACAU
              type: string
            - const: TAIWAN
              type: string
            - const: UNITED_STATES
              type: string
            - const: UNITED_KINGDOM
              type: string
            - const: GERMANY
              type: string
            - const: FRANCE
              type: string
            - const: ITALY
              type: string
            - const: SPAIN
              type: string
            - const: JAPAN
              type: string
            - const: SOUTH_KOREA
              type: string
            - type: "null"
        copyLanguage:
          anyOf:
            - minLength: 1
              maxLength: 64
              type: string
            - type: "null"
        reasoningProviderId:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
        reasoningModelId:
          anyOf:
            - type: string
            - type: "null"
        imageProviderId:
          anyOf:
            - format: uuid
              type: string
            - type: "null"
        imageModelId:
          anyOf:
            - type: string
            - type: "null"
        segmentationModel:
          anyOf:
            - $ref: "#/components/schemas/SegmentationModelRef"
            - type: "null"
        defaultMode:
          type: string
          enum:
            - CREATIVE
            - PIXEL_PROTECTED
        imageResolution:
          $ref: "#/components/schemas/ImageResolution"
        imageAspectRatio:
          $ref: "#/components/schemas/ImageAspectRatio"
        candidatesPerType:
          minimum: 1
          maximum: 4
          type: integer
        webResearchEnabled:
          description: Enable restricted visual-direction web research during Agent planning.
          type: boolean
        archivedAt:
          anyOf:
            - format: date-time
              type: string
            - type: "null"
        planningRevision:
          minimum: 0
          description: Monotonic revision bumped whenever planning-relevant project facts change; planning jobs embed it in their fingerprint so stale plans are never reused.
          type: integer
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        cover:
          $ref: "#/components/schemas/ProjectCover"
...
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          required:
            - assets
            - storyboard
            - items
            - outputs
            - jobs
          properties:
            assets:
              type: array
              items:
                $ref: "#/components/schemas/Asset"
            storyboard:
              anyOf:
                - type: object
                  patternProperties:
                    ^(.*)$: {}
                - type: "null"
            items:
              type: array
              items:
                $ref: "#/components/schemas/StoryboardItem"
            outputs:
              type: array
              items:
                $ref: "#/components/schemas/Output"
            jobs:
              type: array
              items:
                $ref: "#/components/schemas/Job"
...
    SegmentationModelRef:
      description: Segmentation model reference; the referenced model must be declared as a segmentation model on its provider.
      type: object
      required:
        - providerId
        - modelId
      properties:
        providerId:
          format: uuid
          type: string
        modelId:
          type: string
        protocol:
          description: Segmentation API protocol; derived from the model's declared segmentationProtocol and re-stated here for the Worker adapter. A body protocol that contradicts the declared one is rejected.
          type: string
          enum:
            - fal
            - grounded_sam
            - seedream_layerize
            - gitee_sam3
    Storyboard:
      type: object
      required:
        - projectId
        - version
        - status
        - campaignStyleLock
      properties:
        projectId:
          format: uuid
          type: string
        version:
          minimum: 1
          type: integer
        status:
          type: string
          enum:
            - DRAFT
            - CONFIRMED
        campaignStyleLock:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/StoryboardItem"
    StoryboardBundle:
      type: object
      required:
        - storyboard
        - items
      properties:
        storyboard:
          anyOf:
            - type: object
              patternProperties:
                ^(.*)$: {}
            - type: "null"
        items:
          type: array
          items:
            $ref: "#/components/schemas/StoryboardItem"
    StoryboardItem:
      type: object
      required:
        - id
        - assetType
        - displayName
        - candidateCount
        - mode
        - status
        - promptInstruction
        - riskFlags
      properties:
        id:
          format: uuid
          type: string
        assetType:
          description: Immutable ecom-details-image template ID; it cannot be changed after planning.
          type: string
        displayName:
          type: string
        shotRole:
          description: Visual-task semantics assigned during planning; immutable. Null on items planned before this field existed.
          anyOf:
            - $ref: "#/components/schemas/StoryboardShotRole"
            - type: "null"
        templateVariant:
          anyOf:
            - type: string
            - type: "null"
        candidateCount:
          minimum: 1
          maximum: 4
          type: integer
        imageProviderId:
          format: uuid
          type: string
        imageModelId:
          type: string
        imageResolution:
          $ref: "#/components/schemas/ImageResolution"
        imageAspectRatio:
          $ref: "#/components/schemas/ImageAspectRatio"
        referencedAssets:
          type: array
          items:
            format: uuid
            type: string
        mode:
          type: string
          enum:
            - CREATIVE
            - PIXEL_PROTECTED
        status:
          type: string
          enum:
            - DRAFT
            - CONFIRMED
            - GENERATING
            - GENERATED
        promptInstruction:
          description: Final image-generation prompt produced by Pi Agent. Must instruct the image model to preserve exact product identity. Worker prefixes selected image roles in their actual request order before sending it to the image model.
          type: string
        factClaims:
          type: array
          items:
            type: string
        riskFlags:
          type: array
          items:
            type: string
...
    UpdateProjectInput:
      type: object
      properties:
        name:
          minLength: 1
          type: string
        category:
          anyOf:
            - type: string
            - type: "null"
        productDescription:
          anyOf:
            - type: string
            - type: "null"
        verifiedFacts:
          type: array
          items:
            type: string
        prohibitedClaims:
          type: array
          items:
            type: string
        brandGuidelines:
          type: object
          patternProperties:
            ^(.*)$: {}
        platformTargets:
          maxItems: 1
          type: array
          items:
            type: string
            enum:
              - TAOBAO
              - JD
              - PDD
              - DOUYIN
              - AMAZON
              - SHOPIFY
        targetMarket:
          anyOf:
            - const: CHINA_MAINLAND
              type: string
            - const: HONG_KONG
              type: string
            - const: MACAU
              type: string
            - const: TAIWAN
              type: string
            - const: UNITED_STATES
              type: string
            - const: UNITED_KINGDOM
              type: string
            - const: GERMANY
              type: string
            - const: FRANCE
              type: string
            - const: ITALY
              type: string
            - const: SPAIN
              type: string
            - const: JAPAN
              type: string
            - const: SOUTH_KOREA
              type: string
            - type: "null"
        copyLanguage:
          anyOf:
            - minLength: 1
              maxLength: 64
              type: string
            - type: "null"
        reasoningModel:
          $ref: "#/components/schemas/ModelRef"
        imageModel:
          $ref: "#/components/schemas/ModelRef"
        segmentationModel:
          anyOf:
            - $ref: "#/components/schemas/SegmentationModelRef"
            - type: "null"
        defaultMode:
          type: string
          enum:
            - CREATIVE
            - PIXEL_PROTECTED
        imageResolution:
          $ref: "#/components/schemas/ImageResolution"
        imageAspectRatio:
          $ref: "#/components/schemas/ImageAspectRatio"
        candidatesPerType:
          minimum: 1
          maximum: 4
          type: integer
        webResearchEnabled:
          description: Enable restricted visual-direction web research during Agent planning.
          type: boolean
        archived:
          description: Archive or restore the project.
          type: boolean
...
    UpdateStoryboardItemInput:
      type: object
      properties:
        assetType:
          description: Immutable ecom-details-image template ID; it cannot be changed after planning.
          type: string
        displayName:
          type: string
        templateVariant:
          anyOf:
            - type: string
            - type: "null"
        candidateCount:
          minimum: 1
          maximum: 4
          type: integer
        imageModel:
          $ref: "#/components/schemas/ModelRef"
        imageResolution:
          $ref: "#/components/schemas/ImageResolution"
        imageAspectRatio:
          $ref: "#/components/schemas/ImageAspectRatio"
        referencedAssets:
          type: array
          items:
            format: uuid
            type: string
        mode:
          type: string
          enum:
            - CREATIVE
            - PIXEL_PROTECTED
        promptInstruction:
          description: Editable final image-generation prompt. Must keep instructing the image model to preserve exact product identity. Worker prefixes selected image roles in their actual request order before sending it to the image model.
          maxLength: 4000
          type: string
...
    ConfirmStoryboardInput:
      type: object
      required:
        - version
      properties:
        version:
          minimum: 1
          description: The storyboard version the client confirmed; must match the current version or the request is rejected with 409.
          type: integer
...
    CreateGenerationJobInput:
      type: object
      required:
        - storyboardItemIds
      properties:
        storyboardItemIds:
          minItems: 1
          type: array
          items:
            format: uuid
            type: string
        generationBatchId:
          format: uuid
          type: string
        revision:
          type: string
        generationConfig:
          type: object
          properties:
            imageResolution:
              $ref: "#/components/schemas/ImageResolution"
            imageAspectRatio:
              $ref: "#/components/schemas/ImageAspectRatio"
            quality:
              description: 生图质量档位；仅 gpt-image 家族消费。缺省 high。
              $ref: "#/components/schemas/ImageQuality"
            outputFormat:
              description: 产物图片格式；仅支持的模型会收到该参数，缺省 png。
              $ref: "#/components/schemas/ImageOutputFormat"
            candidateCount:
              minimum: 1
              maximum: 4
              type: integer
            imageModel:
              $ref: "#/components/schemas/ModelRef"
...
    GenerationJobsResponse:
      type: object
      required:
        - jobs
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
...
    UploadAssetInput:
      type: object
      required:
        - file
      properties:
        file:
          format: binary
          type: string
        role:
          $ref: "#/components/schemas/AssetRole"
        kind:
          $ref: "#/components/schemas/UserAssetKind"
...